fix: move fulfillment handler so it works
